add reconnect_tube

+++ py/beanstalkm/__init__.py
@@ -202,24 +202,32 @@ class Client(Connection):
         message.send(tube)
         return message
 
-    def reserve(self, timeout=None, drop=True):
+    def reserve(self, timeout=None, drop=True, reconnect_tube=None):
         message = None
-        try:
-            if timeout is None:
-                message = self.queue.reserve()
+        if self.socket is None and self.host is not None and self.port is not None:
+            try:
+                self.reconnect()
+                if reconnect_tube is not None:
+                    self.queue.watch(reconnect_tube)
+            except Exception as e:
+                error_print("Connection error")
+        if self.socket is not None:
+            try:
+                if timeout is None:
+                    message = self.queue.reserve()
+                else:
+                    message = self.queue('reserve_with_timeout', int(timeout))
+            except CommandFailed as e:
+                name, status, result = e.args
+                if status == 'TIMED_OUT':
+                    return None
+                elif status == 'DEADLINE_SOON':
+                    return None
+            except SocketError:
+                self.socket = None
             else:
-                message = self.queue('reserve_with_timeout', int(timeout))
-        except CommandFailed as e:
-            name, status, result = e.args
-            if status == 'TIMED_OUT':
-                return None
-            elif status == 'DEADLINE_SOON':
-                return None
-        except SocketError:
-            self.socket = None
-        else:
-            if drop:
-                message.delete()
+                if drop:
+                    message.delete()
         return message
 
     @catch(message="Delete a job failed: %s")
